fixing all imports & database fetching
