+++ src/database/sample_database.sql
@@ -2,11 +2,11 @@ show dbs
 use lunchbox
 show collections
 // CANTEEN COLLECTION
-db.createCollection("canteen")
+db.createCollection("canteens")
 db.canteen.insert({"name": "Food Point", "owner": "Shah Rukh", "location": "Banashankari", "city": "Bengaluru"})
 db.canteen.insert({"name": "Vidhyarthi Bhavan", "owner": "Basaraj S. N.", "location": "Basavangudi", "city": "Bengaluru"})
 // CATERER COLLECTION
-db.createCollection("caterer")
+db.createCollection("caterers")
 db.caterer.insert({"name": "Delhi Marathi Catering Services", "owner": "Shivaji", "location": "Delhi-6", "city": "Bengaluru"})
 db.caterer.insert({"name": "Vintage Delhi Catering Services", "owner": "Samuel", "location": "Vasant Kunj", "city": "Delhi"})
 // USERS COLLECTION
