+++ src/scripts/institution.py
@@ -1,1 +1,2 @@
 # all endpoints related to institution view
+from flask_app import *
